refactor: split large route components
