+++ src/routes/Connections/index.tsx
@@ -1,37 +1,44 @@
 import type {
   ConnectionAuthType,
-  ConnectionAppSummary,
   ConnectionConnectInput,
-  ConnectionExecutionLogSummary,
   ConnectionProviderDetail,
   ConnectionProviderSummary,
   ConnectionSummary,
-  ConnectionUsageDailyPoint,
-  ConnectionUsageServiceItem,
 } from "../../../electron/connections/common.ts"
 import type { ConnectionErrorNotice } from "./connection-error-display.ts"
+import type { ConnectionCatalogFilter, ConnectionCategoryFilter, DisconnectTarget } from "./connection-route-model.ts"
 import type { UseConnections } from "@/hooks/useConnections"
-import type { MessageKey, TranslateFn } from "@/i18n/i18n"
 import type { UserFacingError } from "@/lib/user-facing-error"
 
-import {
-  AlertCircle,
-  ArrowLeft,
-  BarChart3,
-  ChevronDown,
-  ExternalLink,
-  KeyRound,
-  LoaderCircle,
-  Plug,
-  RefreshCw,
-  Star,
-  Unplug,
-  X,
-} from "lucide-react"
+import { AlertCircle, ArrowLeft, ChevronDown, ExternalLink, KeyRound, Plug, X } from "lucide-react"
 import * as React from "react"
-import { connectionAppDisplayLabel as connectionAppUiDisplayLabel } from "../../../electron/connections/summary.ts"
 import { ConnectDialog } from "./ConnectDialog.tsx"
 import { getConnectionDetailErrorNotice, getConnectionListErrorNotice } from "./connection-error-display.ts"
+import {
+  authTypeNeedsDialog,
+  buildCategoryFilters,
+  categoryFilterLimit,
+  categoryFilterPrefix,
+  detailPaneAnimationMs,
+  formatAuthTypes,
+  formatDateTime,
+  formatProviderCategoryLabels,
+  getDefaultAuthType,
+  getEmptyState,
+  getFilterValue,
+  getProviderActionLabel,
+  getProviderDescription,
+  getProviderMeta,
+  getProviderStatusDisplayLabel,
+  getProviderStatusTone,
+  isConnected,
+  matchesProviderFilter,
+  matchesProviderQuery,
+  parseFilterValue,
+} from "./connection-route-model.ts"
+import { AuthTypeToggleGroup, ConnectionAccountsList } from "./ConnectionAccountsList.tsx"
+import { ProviderUsagePanel } from "./ConnectionUsagePanel.tsx"
+import { DisconnectDialog } from "./DisconnectDialog.tsx"
 import {
   getProviderGridColumnCount,
   getProviderGridVisibleRange,
@@ -39,13 +46,11 @@ import {
   providerGridGapPx,
 } from "./provider-grid-virtualization.ts"
 import { ProviderIcon } from "./ProviderIcon.tsx"
-import { authTypeLabel } from "./shared.ts"
 import { Loader } from "@/components/ai-elements/loader"
 import { ErrorNotice } from "@/components/ErrorNotice"
 import { SearchField } from "@/components/SearchField"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
-import { Dialog } from "@/components/ui/dialog"
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -65,292 +70,8 @@ import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group"
 import { isConnectionPollingTarget } from "@/hooks/connection-oauth-pending"
 import { useT } from "@/i18n/i18n"
 import { resolveConnectionError } from "@/lib/connections-error"
-import { resolveUserFacingError, userFacingErrorDescription } from "@/lib/user-facing-error"
+import { userFacingErrorDescription } from "@/lib/user-facing-error"
 import { cn } from "@/lib/utils"
-
-const executionLogLimit = 12
-const detailPaneAnimationMs = 150
-const categoryFilterLimit = 4
-const accountActionButtonClassName = "h-7 gap-1.5 px-2"
-const categoryFilterPrefix = "category:"
-const uncategorizedCategoryValue = "__uncategorized__"
-const categoryMessageKeysByRawLabel: Record<string, MessageKey> = {
-  AI: "connections.category.ai",
-  Communication: "connections.category.communication",
-  "Data & Analytics": "connections.category.dataAnalytics",
-  "Design & Media": "connections.category.designMedia",
-  "Developer Tools": "connections.category.developerTools",
-  Documentation: "connections.category.documentation",
-  Efficiency: "connections.category.efficiency",
-  Finance: "connections.category.finance",
-  "Maps & Location": "connections.category.mapsLocation",
-  Marketing: "connections.category.marketing",
-  Productivity: "connections.category.productivity",
-  "Security & Identity": "connections.category.securityIdentity",
-  Social: "connections.category.social",
-  Storage: "connections.category.storage",
-}
-
-type ConnectionCatalogFilter =
-  | { kind: "all" }
-  | { kind: "attention" }
-  | { kind: "category"; category: string }
-  | { kind: "connected" }
-
-interface ConnectionCategoryFilter {
-  count: number
-  displayLabel: string
-  label: string
-}
-
-interface DisconnectTarget {
-  app?: ConnectionAppSummary
-  provider: ConnectionProviderSummary
-}
-
-function isConnected(provider: ConnectionProviderSummary): boolean {
-  return provider.status === "connected" && provider.appStatus === "active"
-}
-
-function getProviderStatusTone(provider: ConnectionProviderSummary): "attention" | "available" | "connected" {
-  if (provider.status === "needs_attention") {
-    return "attention"
-  }
-  if (provider.status === "connected") {
-    return "connected"
-  }
-  return "available"
-}
-
-function getProviderStatusLabel(provider: ConnectionProviderSummary, t: ReturnType<typeof useT>): string | null {
-  switch (provider.status) {
-    case "needs_attention":
-      return t("connections.providerNeedsAttention")
-    case "available":
-      return t("connections.providerAvailable")
-    case "connected":
-      return null
-  }
-}
-
-function getDefaultAuthType(provider: ConnectionProviderSummary): Exclude<ConnectionAuthType, null> | null {
-  if (provider.appAuthType && provider.authTypes.includes(provider.appAuthType)) {
-    return provider.appAuthType
-  }
-  return provider.authTypes[0] ?? null
-}
-
-function formatAuthTypes(authTypes: Exclude<ConnectionAuthType, null>[], t: ReturnType<typeof useT>): string {
-  if (authTypes.length === 0) {
-    return t("connections.authUnknown")
-  }
-  return authTypes.map((authType) => authTypeLabel(t, authType)).join(" / ")
-}
-
-function isConnectionAuthType(
-  value: string,
-  authTypes: Exclude<ConnectionAuthType, null>[],
-): value is Exclude<ConnectionAuthType, null> {
-  return authTypes.some((authType) => authType === value)
-}
-
-function formatDateTime(value: number | string | undefined, t: ReturnType<typeof useT>): string {
-  if (!value) {
-    return t("connections.notConnected")
-  }
-
-  const date = typeof value === "number" ? new Date(value) : new Date(value)
-  if (Number.isNaN(date.getTime())) {
-    return t("connections.executionTimeUnknown")
-  }
-
-  return date.toLocaleString([], {
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-    month: "2-digit",
-  })
-}
-
-function formatUsageDate(value: string): string {
-  const date = new Date(`${value}T00:00:00`)
-  if (Number.isNaN(date.getTime())) {
-    return value
-  }
-  return date.toLocaleDateString([], { day: "2-digit", month: "2-digit" })
-}
-
-function formatDuration(durationMs: number | null, t: ReturnType<typeof useT>): string {
-  if (durationMs === null) {
-    return t("connections.executionDurationUnknown")
-  }
-  if (durationMs < 1000) {
-    return t("connections.executionDurationMs", { value: durationMs })
-  }
-  return t("connections.executionDurationSeconds", {
-    value: Intl.NumberFormat(undefined, { maximumFractionDigits: 1 }).format(durationMs / 1000),
-  })
-}
-
-function getProviderDescription(provider: ConnectionProviderSummary, t: ReturnType<typeof useT>): string {
-  switch (provider.status) {
-    case "needs_attention":
-      return t("connections.providerNeedsAttentionDescription", { name: provider.displayName })
-    case "connected":
-      if (provider.appCount > 1) {
-        return t("connections.connectionCount", { count: provider.appCount })
-      }
-      return provider.accountLabel ?? getProviderCategoryLabel(provider, t)
-    case "available":
-      return getProviderCategoryLabel(provider, t)
-  }
-}
-
-function getEmptyState(
-  summary: ConnectionSummary | null,
-  t: ReturnType<typeof useT>,
-): { description: string; title: string } {
-  if (!summary) {
-    return { title: t("connections.unavailableTitle"), description: t("connections.unavailableDescription") }
-  }
-
-  switch (summary.status) {
-    case "signed-out":
-      return { title: t("connections.signedOutTitle"), description: t("connections.signedOutDescription") }
-    case "unavailable":
-      return { title: t("connections.unavailableTitle"), description: t("connections.unavailableDescription") }
-    case "ready":
-      return { title: t("connections.emptyTitle"), description: t("connections.readyEmptyDescription") }
-  }
-}
-
-function authTypeNeedsDialog(authType: Exclude<ConnectionAuthType, null>): boolean {
-  return authType === "api_key" || authType === "custom_credential" || authType === "federated"
-}
-
-function getProviderStatusDisplayLabel(provider: ConnectionProviderSummary, t: ReturnType<typeof useT>): string {
-  return getProviderStatusLabel(provider, t) ?? t("connections.connected")
-}
-
-function getProviderActionLabel(provider: ConnectionProviderSummary, t: ReturnType<typeof useT>): string {
-  if (provider.actionKind === "unavailable") {
-    return t("connections.unsupported")
-  }
-  switch (provider.status) {
-    case "needs_attention":
-      return t("connections.reconnect")
-    case "connected":
-      return t("connections.manage")
-    case "available":
-      return t("connections.connect")
-  }
-}
-
-function getCategoryDisplayLabel(label: string, t: TranslateFn): string {
-  if (label === uncategorizedCategoryValue) {
-    return t("connections.categoryUnknown")
-  }
-  const key = categoryMessageKeysByRawLabel[label]
-  return key ? t(key) : label
-}
-
-function getProviderCategoryRawLabels(provider: ConnectionProviderSummary): string[] {
-  return provider.categoryLabels.length > 0 ? provider.categoryLabels : [uncategorizedCategoryValue]
-}
-
-function getProviderCategoryLabel(provider: ConnectionProviderSummary, t: TranslateFn): string {
-  return getCategoryDisplayLabel(getProviderCategoryRawLabels(provider)[0] ?? uncategorizedCategoryValue, t)
-}
-
-function formatProviderCategoryLabels(provider: ConnectionProviderSummary, t: TranslateFn): string {
-  return getProviderCategoryRawLabels(provider)
-    .map((label) => getCategoryDisplayLabel(label, t))
-    .join(" / ")
-}
-
-function getProviderMeta(provider: ConnectionProviderSummary, t: ReturnType<typeof useT>): string {
-  if (provider.status === "connected" && provider.appCount === 1 && provider.accountLabel) {
-    return provider.accountLabel
-  }
-  if (provider.status === "connected") {
-    return t("connections.connectionCount", { count: provider.appCount })
-  }
-  return getProviderCategoryLabel(provider, t)
-}
-
-function getConnectionAppGeneratedLabel(app: ConnectionAppSummary, index: number, t: ReturnType<typeof useT>): string {
-  const authLabel = app.authType ? authTypeLabel(t, app.authType) : t("connections.authUnknown")
-  return t("connections.generatedConnectionLabel", { auth: authLabel, index: index + 1 })
-}
-
-function getConnectionAppDisplayLabel(app: ConnectionAppSummary, index: number, t: ReturnType<typeof useT>): string {
-  return connectionAppUiDisplayLabel(app) ?? getConnectionAppGeneratedLabel(app, index, t)
-}
-
-function matchesProviderQuery(provider: ConnectionProviderSummary, normalizedQuery: string, t: TranslateFn): boolean {
-  if (!normalizedQuery) {
-    return true
-  }
-  return (
-    provider.displayName.toLowerCase().includes(normalizedQuery) ||
-    provider.service.toLowerCase().includes(normalizedQuery) ||
-    getProviderCategoryRawLabels(provider).some((label) => {
-      return (
-        label.toLowerCase().includes(normalizedQuery) ||
-        getCategoryDisplayLabel(label, t).toLowerCase().includes(normalizedQuery)
-      )
-    }) ||
-    provider.apps.some((app) => {
-      const candidates = [app.displayName, app.alias, app.accountLabel, app.providerAccountId, app.id]
-      return candidates.some((value) => value?.toLowerCase().includes(normalizedQuery))
-    })
-  )
-}
-
-function getFilterValue(filter: ConnectionCatalogFilter): string {
-  return filter.kind === "category" ? `${categoryFilterPrefix}${filter.category}` : filter.kind
-}
-
-function parseFilterValue(value: string): ConnectionCatalogFilter | null {
-  if (value === "all" || value === "connected" || value === "attention") {
-    return { kind: value }
-  }
-  if (value.startsWith(categoryFilterPrefix)) {
-    const category = value.slice(categoryFilterPrefix.length)
-    return category ? { kind: "category", category } : null
-  }
-  return null
-}
-
-function buildCategoryFilters(
-  providers: ConnectionProviderSummary[],
-  t: ReturnType<typeof useT>,
-): ConnectionCategoryFilter[] {
-  const countByCategory = new Map<string, number>()
-  for (const provider of providers) {
-    for (const label of getProviderCategoryRawLabels(provider)) {
-      countByCategory.set(label, (countByCategory.get(label) ?? 0) + 1)
-    }
-  }
-
-  return [...countByCategory.entries()]
-    .map(([label, count]) => ({ count, displayLabel: getCategoryDisplayLabel(label, t), label }))
-    .sort((left, right) => right.count - left.count || left.displayLabel.localeCompare(right.displayLabel))
-}
-
-function matchesProviderFilter(provider: ConnectionProviderSummary, filter: ConnectionCatalogFilter): boolean {
-  switch (filter.kind) {
-    case "all":
-      return true
-    case "connected":
-      return isConnected(provider)
-    case "attention":
-      return provider.status === "needs_attention"
-    case "category":
-      return getProviderCategoryRawLabels(provider).includes(filter.category)
-  }
-}
-
 interface ConnectionsPanelProps {
   authIntent?: ConnectionAuthIntent | null
   connections: UseConnections
@@ -1413,477 +1134,11 @@ function ConnectionPanel({
   )
 }
 
-function ConnectionAccountsList({
-  busy,
-  canSetDefault,
-  connections,
-  onConnect,
-  onDisconnect,
-  polling,
-  provider,
-}: {
-  busy: UseConnections["busy"]
-  canSetDefault: boolean
-  connections: UseConnections
-  onConnect: (
-    provider: ConnectionProviderSummary,
-    authType: Exclude<ConnectionAuthType, null>,
-    appId?: string,
-  ) => Promise<void>
-  onDisconnect: (target: DisconnectTarget) => void
-  polling: string | null
-  provider: ConnectionProviderSummary
-}) {
-  const t = useT()
-  return (
-    <div className="grid gap-2">
-      <div className="flex min-w-0 items-center justify-between gap-2 px-0.5">
-        <h4 className="oo-text-caption font-medium text-foreground">{t("connections.connectionAccounts")}</h4>
-        <span className="oo-text-micro oo-text-muted shrink-0">
-          {t("connections.connectionCount", { count: provider.apps.length })}
-        </span>
-      </div>
-      {provider.apps.map((app, index) => {
-        const isPolling = isConnectionPollingTarget(polling, provider.service, app.id)
-        const reconnectAuthType =
-          app.authType && app.authType !== "no_auth" && isConnectionAuthType(app.authType, provider.authTypes)
-            ? app.authType
-            : null
-        const authLabel = app.authType ? authTypeLabel(t, app.authType) : t("connections.authUnknown")
-        const accountLabel = getConnectionAppDisplayLabel(app, index, t)
-        return (
-          <article
-            key={app.id}
-            className="grid min-w-0 gap-2.5 rounded-md border bg-card px-3 py-2.5 text-card-foreground"
-          >
-            <div className="grid min-w-0 gap-1">
-              <div className="flex min-w-0 flex-wrap items-start gap-1.5">
-                <span className="oo-text-control max-w-full min-w-0 font-medium break-all">{accountLabel}</span>
-                <span className="flex shrink-0 flex-wrap items-center gap-1.5">
-                  {app.isDefault ? <Badge variant="success">{t("connections.defaultConnection")}</Badge> : null}
-                  {app.status === "reauth_required" || app.status === "error" ? (
-                    <Badge variant="warning">{t("connections.providerNeedsAttention")}</Badge>
-                  ) : null}
-                </span>
-              </div>
-              <div className="oo-text-micro oo-text-muted flex min-w-0 flex-wrap items-center gap-1.5">
-                <span className="shrink-0">{authLabel}</span>
-              </div>
-            </div>
-            <div className="flex min-w-0 flex-wrap items-center gap-1.5 border-t pt-2">
-              {canSetDefault && !app.isDefault && provider.apps.length > 1 ? (
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  className={accountActionButtonClassName}
-                  onClick={() => void connections.setDefaultAccount(provider.service, app.id)}
-                >
-                  <Star className="size-3.5" />
-                  {t("connections.setDefaultConnection")}
-                </Button>
-              ) : null}
-              {reconnectAuthType ? (
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  className={accountActionButtonClassName}
-                  disabled={isPolling || busy === "connect"}
-                  onClick={() => void onConnect(provider, reconnectAuthType, app.id)}
-                >
-                  {isPolling ? <Loader size={14} /> : <KeyRound className="size-3.5" />}
-                  {isPolling ? t("connections.oauthWaiting") : t("connections.reconnect")}
-                </Button>
-              ) : null}
-              {provider.canDisconnect ? (
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  disabled={busy === "disconnect"}
-                  className={cn(
-                    accountActionButtonClassName,
-                    "border-[var(--oo-danger-border)] text-destructive hover:bg-[var(--oo-danger-surface)] hover:text-destructive",
-                  )}
-                  onClick={() => onDisconnect({ provider, app })}
-                >
-                  <Unplug className="size-3.5" />
-                  {t("connections.disconnect")}
-                </Button>
-              ) : null}
-            </div>
-          </article>
-        )
-      })}
-    </div>
-  )
-}
-
-function AuthTypeToggleGroup({
-  authTypes,
-  onChange,
-  value,
-}: {
-  authTypes: Exclude<ConnectionAuthType, null>[]
-  onChange: (value: Exclude<ConnectionAuthType, null>) => void
-  value: Exclude<ConnectionAuthType, null> | null
-}) {
-  const t = useT()
-
-  if (authTypes.length <= 1) {
-    return null
-  }
-
-  return (
-    <ToggleGroup
-      variant="outline"
-      size="sm"
-      type="single"
-      value={value ?? undefined}
-      aria-label={t("connections.authMode")}
-      onValueChange={(nextValue) => {
-        if (isConnectionAuthType(nextValue, authTypes)) {
-          onChange(nextValue)
-        }
-      }}
-    >
-      {authTypes.map((authType) => (
-        <ToggleGroupItem key={authType} value={authType}>
-          {authTypeLabel(t, authType)}
-        </ToggleGroupItem>
-      ))}
-    </ToggleGroup>
-  )
-}
-
-function ProviderUsagePanel({
-  connections,
-  provider,
-  usage,
-  usageDays,
-}: {
-  connections: UseConnections
-  provider: ConnectionProviderSummary
-  usage?: ConnectionUsageServiceItem
-  usageDays: number
-}) {
-  const t = useT()
-  const [isUsageDialogOpen, setIsUsageDialogOpen] = React.useState(false)
-  const [logs, setLogs] = React.useState<ConnectionExecutionLogSummary | null>(null)
-  const [loading, setLoading] = React.useState(false)
-  const [error, setError] = React.useState<UserFacingError | null>(null)
-  const providerUsage = usage ?? {
-    calls: 0,
-    errors: 0,
-    recent: null,
-    service: provider.service,
-    success: 0,
-    trend: [],
-  }
-
-  const loadLogs = React.useCallback(async () => {
-    if (providerUsage.calls <= 0) {
-      return
-    }
-    setLoading(true)
-    setError(null)
-    try {
-      setLogs(await connections.getExecutionLogs({ service: provider.service, limit: executionLogLimit }))
-    } catch (err) {
-      setError(resolveUserFacingError(err, { area: "connections" }))
-    } finally {
-      setLoading(false)
-    }
-  }, [connections, provider.service, providerUsage.calls])
-
-  React.useEffect(() => {
-    setIsUsageDialogOpen(false)
-    setLogs(null)
-    setError(null)
-  }, [provider.service])
-
-  React.useEffect(() => {
-    if (isUsageDialogOpen && providerUsage.calls > 0 && !logs && !loading) {
-      void loadLogs()
-    }
-  }, [isUsageDialogOpen, loadLogs, loading, logs, providerUsage.calls])
-
-  const usageTitle =
-    providerUsage.calls > 0
-      ? t("connections.usageCompactCalls", { count: providerUsage.calls })
-      : t("connections.usageNoCalls")
-  const usageDetail =
-    providerUsage.calls <= 0
-      ? t("connections.usageCompactEmpty", { days: usageDays })
-      : providerUsage.errors > 0
-        ? t("connections.usageCompactWithIncomplete", {
-            days: usageDays,
-            errors: providerUsage.errors,
-            success: providerUsage.success,
-          })
-        : t("connections.usageCompactAllSuccess", { days: usageDays })
-
-  return (
-    <section className="grid gap-1.5">
-      <div className="flex min-w-0 items-center justify-between gap-2 px-0.5">
-        <h3 className="oo-text-title truncate">{t("connections.usageTitle")}</h3>
-        <Button
-          type="button"
-          variant="ghost"
-          size="sm"
-          className="shrink-0 gap-1.5"
-          aria-label={t("connections.viewUsageForProvider", { name: provider.displayName })}
-          onClick={() => setIsUsageDialogOpen(true)}
-        >
-          <BarChart3 className="size-3.5" />
-          {t("connections.viewUsage")}
-        </Button>
-      </div>
-      <div className="grid min-w-0 gap-2 rounded-md bg-[var(--oo-inspector-surface)] px-2.5 py-2">
-        <div className="flex min-w-0 items-start justify-between gap-3">
-          <div className="grid min-w-0 gap-0.5">
-            <div className="oo-text-title truncate">{usageTitle}</div>
-            <div className="oo-text-caption oo-text-muted truncate">{usageDetail}</div>
-          </div>
-        </div>
-      </div>
-      <UsageDialog
-        open={isUsageDialogOpen}
-        provider={provider}
-        usage={providerUsage}
-        usageDays={usageDays}
-        logs={logs}
-        loading={loading}
-        error={error}
-        onClose={() => setIsUsageDialogOpen(false)}
-        onRefresh={() => void loadLogs()}
-      />
-    </section>
-  )
-}
-
-function UsageDialog({
-  open,
-  provider,
-  usage,
-  usageDays,
-  logs,
-  loading,
-  error,
-  onClose,
-  onRefresh,
-}: {
-  error: UserFacingError | null
-  loading: boolean
-  logs: ConnectionExecutionLogSummary | null
-  onClose: () => void
-  onRefresh: () => void
-  open: boolean
-  provider: ConnectionProviderSummary
-  usage: ConnectionUsageServiceItem
-  usageDays: number
-}) {
-  const t = useT()
-  const successRate = usage.calls > 0 ? Math.round((usage.success / usage.calls) * 100) : 0
-
-  return (
-    <Dialog
-      open={open}
-      onClose={onClose}
-      closeLabel={t("common.cancel")}
-      title={t("connections.usageDialogTitle")}
-      description={t("connections.usageDialogDescription", {
-        days: usageDays,
-        name: provider.displayName,
-      })}
-      className="max-w-2xl"
-    >
-      <div className="grid gap-3">
-        <div className="flex min-w-0 flex-wrap items-center justify-between gap-2">
-          <Badge variant="outline">{t("connections.usageRangeLabel", { days: usageDays })}</Badge>
-          <Button variant="ghost" size="sm" disabled={loading || usage.calls <= 0} onClick={onRefresh}>
-            <RefreshCw className={cn("size-4", loading && "animate-spin")} />
-            {t("connections.refreshExecutionLogs")}
-          </Button>
-        </div>
-        <div className="grid grid-cols-4 gap-2 max-[640px]:grid-cols-2">
-          <Metric label={t("connections.usageCalls")} value={usage.calls} />
-          <Metric label={t("connections.usageSuccess")} value={usage.success} />
-          <Metric label={t("connections.usageIncomplete")} value={usage.errors} />
-          <Metric label={t("connections.usageSuccessRate")} value={`${successRate}%`} />
-        </div>
-        <UsageTrend points={usage.trend} />
-        <ExecutionLogs logs={logs} loading={loading} error={error} />
-      </div>
-    </Dialog>
-  )
-}
-
-function Metric({ label, value }: { label: string; value: number | string }) {
-  return (
-    <div className="rounded-md bg-muted/45 px-2.5 py-2">
-      <div className="oo-text-title">{value}</div>
-      <div className="oo-text-caption oo-text-muted">{label}</div>
-    </div>
-  )
-}
-
-function UsageTrend({ points }: { points: ConnectionUsageDailyPoint[] }) {
-  const t = useT()
-  const sorted = [...points].sort((left, right) => left.date.localeCompare(right.date))
-  const maxCalls = Math.max(1, ...sorted.map((point) => point.calls))
-
-  if (sorted.every((point) => point.calls === 0)) {
-    return (
-      <div className="grid h-24 place-items-center rounded-md bg-muted/35">
-        <span className="oo-text-caption oo-text-muted">{t("connections.usageDailyUnavailable")}</span>
-      </div>
-    )
-  }
-
-  return (
-    <div className="grid h-28 min-w-0 grid-rows-[minmax(0,1fr)_auto] gap-2 rounded-md bg-muted/35 px-3 py-3">
-      <div className="flex min-w-0 items-end gap-1.5">
-        {sorted.map((point) => {
-          const height = point.calls > 0 ? Math.max(8, Math.round((point.calls / maxCalls) * 100)) : 2
-          return (
-            <div
-              key={point.date}
-              className="flex h-full min-w-0 flex-1 items-end"
-              title={`${point.date}: ${point.calls}`}
-            >
-              <div
-                className={cn("w-full rounded-t-sm bg-muted-foreground/45", point.errors > 0 && "bg-destructive/70")}
-                style={{ height: `${height}%` }}
-              />
-            </div>
-          )
-        })}
-      </div>
-      <div className="grid auto-cols-fr grid-flow-col gap-1.5">
-        {sorted.map((point) => (
-          <span key={point.date} className="oo-text-micro oo-text-muted truncate text-center">
-            {formatUsageDate(point.date)}
-          </span>
-        ))}
-      </div>
-    </div>
-  )
-}
-
-function ExecutionLogs({
-  logs,
-  loading,
-  error,
-}: {
-  logs: ConnectionExecutionLogSummary | null
-  loading: boolean
-  error: UserFacingError | null
-}) {
-  const t = useT()
-  return (
-    <div className="grid gap-2">
-      <div className="flex min-w-0 items-center gap-2">
-        <BarChart3 className="oo-icon-muted size-4" />
-        <h4 className="oo-text-label">{t("connections.executionLogsTitle")}</h4>
-      </div>
-      {loading ? (
-        <div className="grid gap-1">
-          <div className="h-9 animate-pulse rounded-md bg-muted" />
-          <div className="h-9 animate-pulse rounded-md bg-muted" />
-          <div className="h-9 animate-pulse rounded-md bg-muted" />
-        </div>
-      ) : error ? (
-        <ErrorNotice error={error} compact />
-      ) : !logs || logs.items.length === 0 ? (
-        <div className="oo-text-caption oo-text-muted rounded-md bg-muted/35 px-3 py-2">{t("connections.noData")}</div>
-      ) : (
-        <div className="overflow-hidden rounded-md border">
-          {logs.items.map((item) => (
-            <div
-              key={item.id}
-              className="grid grid-cols-[minmax(0,1fr)_auto] gap-2 border-b px-2.5 py-2 last:border-b-0"
-            >
-              <div className="min-w-0">
-                <div className="oo-text-control truncate font-mono">{item.action}</div>
-                <div className="oo-text-micro oo-text-muted truncate">
-                  {formatDateTime(item.finishedAt, t)} · {formatDuration(item.durationMs, t)}
-                </div>
-              </div>
-              <Badge variant={item.status === "success" ? "success" : "warning"}>
-                {item.status === "success"
-                  ? t("connections.executionStatusSuccess")
-                  : t("connections.executionStatusError")}
-              </Badge>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  )
-}
-
 function DetailRow({ label, value, mono }: { label: string; value: string; mono?: boolean }) {
   return (
     <div className="grid min-w-0 grid-cols-[6.5rem_minmax(0,1fr)] border-b px-2.5 py-1.5 last:border-b-0">
       <dt className="oo-text-caption oo-text-muted">{label}</dt>
       <dd className={cn("oo-text-control min-w-0 truncate", mono && "font-mono")}>{value}</dd>
     </div>
-  )
-}
-
-function DisconnectDialog({
-  target,
-  busy,
-  onClose,
-  onConfirm,
-}: {
-  target: DisconnectTarget | null
-  busy: boolean
-  onClose: () => void
-  onConfirm: (target: DisconnectTarget) => void
-}) {
-  const t = useT()
-  if (!target) {
-    return null
-  }
-
-  const { app, provider } = target
-  const displayName = provider.displayName
-  const appIndex = app
-    ? Math.max(
-        0,
-        provider.apps.findIndex((item) => item.id === app.id),
-      )
-    : 0
-  const accountTypeLabel = app ? getConnectionAppDisplayLabel(app, appIndex, t) : t("connections.connectionAccounts")
-
-  return (
-    <Dialog
-      open
-      onClose={busy ? () => undefined : onClose}
-      title={t("connections.confirmDisconnectTitle")}
-      description={t("connections.confirmDisconnectDescription", { name: displayName })}
-      footer={
-        <>
-          <Button variant="outline" disabled={busy} onClick={onClose}>
-            {t("connections.confirmDisconnectCancel")}
-          </Button>
-          <Button variant="outline" disabled={busy} className="text-destructive" onClick={() => onConfirm(target)}>
-            {busy ? <LoaderCircle className="size-4 animate-spin" /> : <Unplug className="size-4" />}
-            {busy ? t("connections.disconnecting") : t("connections.disconnect")}
-          </Button>
-        </>
-      }
-    >
-      <div className="flex min-w-0 items-center gap-2 rounded-lg border px-3 py-2">
-        <ProviderIcon iconUrl={provider.iconUrl} displayName={provider.displayName} />
-        <div className="min-w-0">
-          <div className="oo-text-label truncate">{provider.displayName}</div>
-          <div className="oo-text-caption oo-text-muted truncate">{accountTypeLabel}</div>
-        </div>
-      </div>
-    </Dialog>
   )
 }
