+++ src/routes/Skills/index.tsx
@@ -12,6 +12,7 @@ import type {
   SkillSelectionKey,
   SkillVersionCheckByKey,
 } from "./skill-route-model.ts"
+import type { OrganizationSkillFilter } from "./SkillPageHeader.tsx"
 import type { ObjectStatusTone } from "@/components/ObjectRow"
 import type { UseOrganizationSkills } from "@/hooks/useOrganizationSkills"
 import type { UseOrganizationWorkspace } from "@/hooks/useOrganizationWorkspace"
@@ -19,6 +20,8 @@ import type { TranslateFn as TFunction } from "@/i18n"
 
 import * as React from "react"
 import { toast } from "sonner"
+import { DiscoverSkillsPane } from "./DiscoverSkillsPane.tsx"
+import { InstalledSkillsPane } from "./InstalledSkillsPane.tsx"
 import { planOrganizationSkillBulkLinks } from "./organization-management-model.ts"
 import { useProviderSkillPackageLookup } from "./provider-skill-package-lookup.ts"
 import {
@@ -26,18 +29,11 @@ import {
   getInstallableProviderSkillRecommendations,
 } from "./provider-skill-recommendations.ts"
 import {
-  canInstallPublicSkill,
-  formatPublicPackageUpdateTime,
   getGroupRowPackageLine,
   getGroupStatus,
   getLocalSkillPublishPath,
   getOrganizationSkillRuntimeStatus,
-  getPublicPackageInstallState,
-  getPublicPackageMaintainerLine,
-  getPublicPackageMetaLine,
-  getPublicPackagePrimaryInstallSkill,
   getPublicPackagePrimarySkill,
-  getPublicSkillInstallKey,
   getPublicSkillInstallStateLabel,
   getRuntimeHosts,
   getSkillDocumentRootPath,
@@ -48,12 +44,7 @@ import {
   getStatusBadgeClassName,
   hasSkillUpdateAvailable,
   initialPublicPackageCatalogState,
-  isDiscoverSkillFilter,
-  isEmojiIcon,
-  isImageIcon,
-  isInstalledSkillFilter,
   isInstalledSkillGroup,
-  isNearScrollBottom,
   isPublishableLocalSkill,
   matchesInstalledSkillFilter,
   matchesPublicPackageQuery,
@@ -62,6 +53,9 @@ import {
   shouldUpdatePublishedSkill,
   skillDocumentPreviewSource,
 } from "./skill-route-model.ts"
+import { SkillErrorNotice } from "./SkillErrorNotice.tsx"
+import { SkillPageHeader } from "./SkillPageHeader.tsx"
+import { SkillIconFrame, SkillManagementSheet } from "./SkillUiParts.tsx"
 import { MessageResponse } from "@/components/ai-elements/message"
 import { useSkillService } from "@/components/AppContext"
 import {
@@ -76,13 +70,9 @@ import { ErrorNotice } from "@/components/ErrorNotice"
 import { InspectorCard, InspectorInsetCard } from "@/components/InspectorPanel"
 import { ObjectRowSkeletonGroup, SkeletonText } from "@/components/LoadingSkeletons"
 import { ObjectStatusIcon } from "@/components/ObjectRow"
-import { SearchField } from "@/components/SearchField"
-import { normalizeSkillIconSource } from "@/components/skill-icon-source.ts"
-import { SkillIcon } from "@/components/SkillIcon"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
+import { CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Skeleton } from "@/components/ui/skeleton"
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group"
 import { useSkillObjectActions } from "@/components/useSkillObjectActions"
@@ -90,13 +80,6 @@ import { useAppI18n } from "@/i18n"
 import { listMyPublishedSkillPackages, listPublicSkillPackages } from "@/lib/skills-catalog-client"
 import { resolveUserFacingError, userFacingErrorDescription } from "@/lib/user-facing-error"
 import { cn } from "@/lib/utils"
-
-function SkillErrorNotice({ className, error }: { className?: string; error: string | null | undefined }) {
-  if (!error) {
-    return null
-  }
-  return <ErrorNotice error={resolveUserFacingError(error, { area: "skills" })} compact className={className} />
-}
 
 function skillErrorMessage(cause: unknown, t: TFunction): string {
   return userFacingErrorDescription(resolveUserFacingError(cause, { area: "skills" }), t)
@@ -106,12 +89,6 @@ type SkillOperationError = {
   message: string
   operation: "publish" | "update"
   skillId: SkillSelectionKey
-}
-
-type OrganizationSkillFilter = "all" | "recommended" | "configured"
-
-function isOrganizationSkillFilter(value: string): value is OrganizationSkillFilter {
-  return value === "all" || value === "recommended" || value === "configured"
 }
 
 function visibleSkillOperationError(
@@ -857,7 +834,7 @@ export function SkillsRoute({
           <InstalledSkillsPane
             cliUpdateError={cliUpdateError}
             cliVersionCheck={versionResource.data?.cli}
-            detailContentProps={detailContentProps}
+            detailContent={<SkillDetailContent {...detailContentProps} />}
             groups={filteredInstalledGroups}
             isExecutingCliUpdate={isExecutingCliUpdate}
             isDetailOpen={narrowPane === "detail"}
@@ -1845,1071 +1822,6 @@ function SkillDetailContent({
   }
 
   return <div className="oo-text-body oo-text-muted p-4">{t("skills.detailPlaceholder")}</div>
-}
-
-interface SkillPageHeaderProps {
-  activeTab: SkillPageTab
-  discoveryFilter: DiscoverSkillFilter
-  discoveryQuery: string
-  installedFilter: InstalledSkillFilter
-  installedQuery: string
-  organizationFilter: OrganizationSkillFilter
-  organizationQuery: string
-  organizationTabAvailable: boolean
-  onDiscoveryFilterChange: (filter: DiscoverSkillFilter) => void
-  onDiscoveryQueryChange: (value: string) => void
-  onInstalledFilterChange: (filter: InstalledSkillFilter) => void
-  onInstalledQueryChange: (value: string) => void
-  onOrganizationFilterChange: (filter: OrganizationSkillFilter) => void
-  onOrganizationQueryChange: (value: string) => void
-  onTabChange: (tab: SkillPageTab) => void
-}
-
-function SkillPageHeader({
-  activeTab,
-  discoveryFilter,
-  discoveryQuery,
-  installedFilter,
-  installedQuery,
-  organizationFilter,
-  organizationQuery,
-  organizationTabAvailable,
-  onDiscoveryFilterChange,
-  onDiscoveryQueryChange,
-  onInstalledFilterChange,
-  onInstalledQueryChange,
-  onOrganizationFilterChange,
-  onOrganizationQueryChange,
-  onTabChange,
-}: SkillPageHeaderProps) {
-  const { t } = useAppI18n()
-  const isDiscoverTab = activeTab === "discover"
-  const isOrganizationTab = activeTab === "organization"
-  const searchValue = isDiscoverTab ? discoveryQuery : isOrganizationTab ? organizationQuery : installedQuery
-  const searchPlaceholder = isDiscoverTab
-    ? "skills.discoverSearch"
-    : isOrganizationTab
-      ? "skills.organizationSearch"
-      : "skills.installedSearch"
-  const filterValue = isDiscoverTab ? discoveryFilter : isOrganizationTab ? organizationFilter : installedFilter
-  const filterOptions = isDiscoverTab
-    ? [
-        { label: t("skills.discoverFilter.all"), value: "all" },
-        { label: t("skills.discoverFilter.mine"), value: "mine" },
-      ]
-    : isOrganizationTab
-      ? [
-          { label: t("skills.organizationFilter.all"), value: "all" },
-          { label: t("organizations.skillManageRecommended"), value: "recommended" },
-          { label: t("organizations.skillManageConfigured"), value: "configured" },
-        ]
-      : [
-          { label: t("skills.installedFilter.all"), value: "all" },
-          { label: t("skills.installedFilter.updates"), value: "updates" },
-          { label: t("skills.installedFilter.local"), value: "local" },
-        ]
-
-  return (
-    <header className="oo-border-divider flex min-h-12 items-center gap-2 border-b px-3 py-2">
-      <SkillTabList
-        activeTab={activeTab}
-        organizationTabAvailable={organizationTabAvailable}
-        onTabChange={onTabChange}
-      />
-      <div className="grid min-w-0 flex-1 grid-cols-[minmax(0,1fr)_auto] items-center gap-2">
-        <SearchField
-          placeholder={t(searchPlaceholder)}
-          value={searchValue}
-          onChange={(event) => {
-            const value = event.currentTarget.value
-            if (isDiscoverTab) {
-              onDiscoveryQueryChange(value)
-              return
-            }
-            if (isOrganizationTab) {
-              onOrganizationQueryChange(value)
-              return
-            }
-            onInstalledQueryChange(value)
-          }}
-        />
-        <SkillFilterDropdown
-          ariaLabel={t("skills.filter")}
-          options={filterOptions}
-          value={filterValue}
-          onValueChange={(value) => {
-            if (isDiscoverTab && isDiscoverSkillFilter(value)) {
-              onDiscoveryFilterChange(value)
-              return
-            }
-
-            if (isOrganizationTab && isOrganizationSkillFilter(value)) {
-              onOrganizationFilterChange(value)
-              return
-            }
-
-            if (!isDiscoverTab && !isOrganizationTab && isInstalledSkillFilter(value)) {
-              onInstalledFilterChange(value)
-            }
-          }}
-        />
-      </div>
-    </header>
-  )
-}
-
-function SkillTabList({
-  activeTab,
-  organizationTabAvailable,
-  onTabChange,
-}: {
-  activeTab: SkillPageTab
-  organizationTabAvailable: boolean
-  onTabChange: (tab: SkillPageTab) => void
-}) {
-  const { t } = useAppI18n()
-  const tabs: Array<{ label: string; value: SkillPageTab }> = [
-    { label: t("skills.tab.discover"), value: "discover" },
-    { label: t("skills.tab.installed"), value: "installed" },
-    ...(organizationTabAvailable
-      ? [{ label: t("skills.tab.organization"), value: "organization" as SkillPageTab }]
-      : []),
-  ]
-
-  return (
-    <ToggleGroup
-      type="single"
-      variant="outline"
-      size="sm"
-      className="shrink-0"
-      value={activeTab}
-      onValueChange={(value) => {
-        if (value === "organization" || value === "discover" || value === "installed") {
-          onTabChange(value)
-        }
-      }}
-    >
-      {tabs.map((tab) => (
-        <ToggleGroupItem key={tab.value} value={tab.value}>
-          {tab.label}
-        </ToggleGroupItem>
-      ))}
-    </ToggleGroup>
-  )
-}
-
-interface SkillFilterDropdownProps {
-  ariaLabel: string
-  onValueChange: (value: string) => void
-  options: { label: string; value: string }[]
-  value: string
-}
-
-function SkillFilterDropdown({ ariaLabel, onValueChange, options, value }: SkillFilterDropdownProps) {
-  const selectedOption = options.find((option) => option.value === value) ?? options[0]
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          type="button"
-          variant="outline"
-          size="sm"
-          className="max-w-36 min-w-24 justify-between px-2"
-          aria-label={ariaLabel}
-        >
-          <AppIcons.action.settings className="size-3.5" />
-          <span className="min-w-0 truncate">{selectedOption?.label ?? value}</span>
-          <AppIcons.status.disclosure className="size-3.5 text-muted-foreground" />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" sideOffset={6} className="w-36">
-        {options.map((option) => {
-          const selected = option.value === value
-
-          return (
-            <DropdownMenuItem
-              key={option.value}
-              className="min-w-0 justify-between gap-3"
-              aria-checked={selected}
-              onSelect={() => onValueChange(option.value)}
-            >
-              <span className="min-w-0 truncate">{option.label}</span>
-              {selected ? <AppIcons.status.check className="size-4" /> : null}
-            </DropdownMenuItem>
-          )
-        })}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
-
-interface DiscoverSkillsPaneProps {
-  error: string | null
-  filter: DiscoverSkillFilter
-  groupById: ManagedSkillGroupById
-  installingKey: string | null
-  isLoading: boolean
-  isLoadingMore: boolean
-  isSignedIn: boolean
-  locale: string
-  next: string | null
-  onClosePackage: () => void
-  onInstall: (pkg: PublicSkillPackage, skillName?: string) => void
-  onLoadMore: () => void
-  onOpenManagedSkill: (skillName: string) => void
-  onOpenOrganizationRecommendations?: () => void
-  onRetry: () => void
-  onSelectPackage: (pkg: PublicSkillPackage) => void
-  packages: PublicSkillPackage[]
-  providerRecommendations: ProviderSkillRecommendation[]
-  selectedPackage: PublicSkillPackage | undefined
-}
-
-function ProviderSkillRecommendationNotice({
-  count,
-  onOpenOrganizationRecommendations,
-}: {
-  count: number
-  onOpenOrganizationRecommendations: () => void
-}) {
-  const { t } = useAppI18n()
-
-  return (
-    <div className="flex min-w-0 items-center justify-between gap-3 rounded-md border bg-muted/20 px-3 py-2">
-      <div className="grid min-w-0 gap-0.5">
-        <div className="oo-text-label min-w-0 truncate">
-          {t("skills.providerRecommendationsTitle")}
-          <span className="font-normal text-muted-foreground">
-            {" · "}
-            {t("skills.providerRecommendationsCount", { count })}
-          </span>
-        </div>
-        <div className="oo-text-caption-compact min-w-0 truncate text-muted-foreground">
-          {t("skills.providerRecommendationsMarketHint")}
-        </div>
-      </div>
-      <Button
-        type="button"
-        variant="outline"
-        size="sm"
-        className="shrink-0"
-        onClick={onOpenOrganizationRecommendations}
-      >
-        {t("skills.providerRecommendationsOpenOrganization")}
-      </Button>
-    </div>
-  )
-}
-
-function DiscoverSkillsPane({
-  error,
-  filter,
-  groupById,
-  installingKey,
-  isLoading,
-  isLoadingMore,
-  isSignedIn,
-  locale,
-  next,
-  onClosePackage,
-  onInstall,
-  onLoadMore,
-  onOpenManagedSkill,
-  onOpenOrganizationRecommendations,
-  onRetry,
-  onSelectPackage,
-  packages,
-  providerRecommendations,
-  selectedPackage,
-}: DiscoverSkillsPaneProps) {
-  const { t } = useAppI18n()
-  const autoLoadRequestedRef = React.useRef(false)
-  const canLoadMore = Boolean(next) && !isLoading && !isLoadingMore && packages.length > 0
-
-  React.useEffect(() => {
-    if (!isLoadingMore) {
-      autoLoadRequestedRef.current = false
-    }
-  }, [isLoadingMore, next])
-
-  const handleScroll = React.useCallback(
-    (event: React.UIEvent<HTMLDivElement>) => {
-      if (!canLoadMore || autoLoadRequestedRef.current || !isNearScrollBottom(event.currentTarget)) {
-        return
-      }
-
-      autoLoadRequestedRef.current = true
-      onLoadMore()
-    },
-    [canLoadMore, onLoadMore],
-  )
-
-  return (
-    <div className="min-h-0 overflow-auto px-3 py-3" onScroll={handleScroll}>
-      <div className="grid gap-3 pr-1">
-        {onOpenOrganizationRecommendations && providerRecommendations.length > 0 ? (
-          <ProviderSkillRecommendationNotice
-            count={providerRecommendations.length}
-            onOpenOrganizationRecommendations={onOpenOrganizationRecommendations}
-          />
-        ) : null}
-        {error ? (
-          <div className="flex min-w-0 items-start gap-2">
-            <SkillErrorNotice className="min-w-0 flex-1" error={error} />
-            <Button type="button" variant="outline" size="sm" disabled={isLoading} onClick={onRetry}>
-              {isLoading ? (
-                <AppIcons.status.loading className="size-3.5 animate-spin" />
-              ) : (
-                <AppIcons.action.refresh className="size-3.5" />
-              )}
-              {t("skills.retry")}
-            </Button>
-          </div>
-        ) : null}
-        {isLoading && packages.length === 0 ? (
-          <PublicSkillGridSkeleton />
-        ) : packages.length === 0 ? (
-          <div className="oo-text-body oo-text-muted px-1 py-3">
-            {filter === "mine"
-              ? isSignedIn
-                ? t("skills.discoverMineEmpty")
-                : t("skills.discoverMineSignedOut")
-              : t("skills.discoverEmpty")}
-          </div>
-        ) : (
-          <div className="grid grid-cols-[repeat(auto-fill,minmax(15.5rem,1fr))] gap-2.5">
-            {packages.map((pkg) => (
-              <PublicSkillPackageCard
-                key={pkg.id}
-                groupById={groupById}
-                installingKey={installingKey}
-                pkg={pkg}
-                selected={selectedPackage?.id === pkg.id}
-                onInstall={(skillName) => onInstall(pkg, skillName)}
-                onOpenManagedSkill={onOpenManagedSkill}
-                onSelect={() => onSelectPackage(pkg)}
-              />
-            ))}
-          </div>
-        )}
-        {next ? (
-          <div className="flex justify-center py-2">
-            <Button
-              type="button"
-              variant="outline"
-              size="sm"
-              disabled={isLoading || isLoadingMore}
-              onClick={onLoadMore}
-            >
-              {isLoadingMore ? <AppIcons.status.loading className="animate-spin" /> : null}
-              {isLoadingMore ? t("skills.discoverLoadingMore") : t("skills.discoverLoadMore")}
-            </Button>
-          </div>
-        ) : null}
-      </div>
-
-      {selectedPackage ? (
-        <PublicSkillPackageSheet
-          installingKey={installingKey}
-          groupById={groupById}
-          locale={locale}
-          pkg={selectedPackage}
-          onClose={onClosePackage}
-          onInstall={onInstall}
-          onOpenManagedSkill={onOpenManagedSkill}
-        />
-      ) : null}
-    </div>
-  )
-}
-
-function PublicSkillGridSkeleton() {
-  return (
-    <div className="grid grid-cols-[repeat(auto-fill,minmax(15.5rem,1fr))] gap-2.5">
-      {Array.from({ length: 6 }).map((_, index) => (
-        <div key={index} className="grid gap-3 rounded-md border bg-card px-3 py-3">
-          <div className="flex items-start gap-3">
-            <Skeleton className="size-10 rounded-md" />
-            <div className="grid flex-1 gap-2">
-              <SkeletonText className="h-4 w-28" />
-              <SkeletonText className="h-3 w-full" />
-              <SkeletonText className="h-3 w-20" />
-            </div>
-          </div>
-          <Skeleton className="h-8 rounded-md" />
-        </div>
-      ))}
-    </div>
-  )
-}
-
-interface PublicSkillPackageCardProps {
-  groupById: ManagedSkillGroupById
-  installingKey: string | null
-  onInstall: (skillName?: string) => void
-  onOpenManagedSkill: (skillName: string) => void
-  onSelect: () => void
-  pkg: PublicSkillPackage
-  selected: boolean
-}
-
-function PublicSkillPackageCard({
-  groupById,
-  installingKey,
-  onInstall,
-  onOpenManagedSkill,
-  onSelect,
-  pkg,
-  selected,
-}: PublicSkillPackageCardProps) {
-  const { t } = useAppI18n()
-  const primarySkill = getPublicPackagePrimarySkill(pkg)
-  const primaryInstallSkill = getPublicPackagePrimaryInstallSkill(groupById, pkg)
-  const state = getPublicPackageInstallState(groupById, pkg)
-  const isInstalling = installingKey === getPublicSkillInstallKey(pkg, primaryInstallSkill?.name)
-
-  return (
-    <div
-      className={cn(
-        "grid min-h-44 grid-rows-[minmax(0,1fr)_auto] overflow-hidden rounded-md border bg-card text-card-foreground transition-colors hover:bg-[var(--oo-row-hover)]",
-        selected && "border-[var(--accent-ring)] bg-[var(--oo-row-selected)] hover:bg-[var(--oo-row-selected)]",
-      )}
-    >
-      <button
-        type="button"
-        className="grid min-w-0 gap-2 p-3 text-left outline-none focus-visible:ring-[3px] focus-visible:ring-ring/40"
-        onClick={onSelect}
-      >
-        <div className="flex min-w-0 items-start gap-3">
-          <SkillIconFrame icon={pkg.icon} />
-          <div className="grid min-w-0 gap-1">
-            <div className="oo-text-label min-w-0 truncate">{pkg.displayName}</div>
-            <div className="oo-text-caption oo-text-muted min-w-0 truncate" title={pkg.name}>
-              {pkg.name}
-            </div>
-          </div>
-        </div>
-        {pkg.description ? <p className="oo-text-caption line-clamp-2 text-foreground/75">{pkg.description}</p> : null}
-        <div className="oo-text-caption oo-text-muted min-w-0 truncate" title={getPublicPackageMetaLine(pkg, t)}>
-          {getPublicPackageMetaLine(pkg, t)}
-        </div>
-      </button>
-      <div className="oo-border-divider flex items-center justify-between gap-2 border-t px-3 py-2">
-        <div className="min-w-0">
-          {state === "name-conflict" || state === "unavailable" ? (
-            <Badge variant="outline">{getPublicSkillInstallStateLabel(state, t)}</Badge>
-          ) : null}
-        </div>
-        {state === "name-conflict" && primarySkill ? (
-          <Button type="button" variant="ghost" size="sm" onClick={() => onOpenManagedSkill(primarySkill.name)}>
-            {t("skills.discoverOpenManage")}
-          </Button>
-        ) : state === "installed" && primarySkill ? (
-          <Button type="button" variant="ghost" size="sm" onClick={() => onOpenManagedSkill(primarySkill.name)}>
-            {t("skills.installedManage")}
-          </Button>
-        ) : canInstallPublicSkill(state) ? (
-          <Button
-            type="button"
-            variant="outline"
-            size="sm"
-            disabled={isInstalling}
-            onClick={() => onInstall(primaryInstallSkill?.name)}
-          >
-            {isInstalling ? <AppIcons.status.loading className="animate-spin" /> : <AppIcons.action.installPackage />}
-            {isInstalling
-              ? t("skills.registryInstalling")
-              : state === "partially-installed"
-                ? t("skills.discoverInstallMissing")
-                : t("organizations.skillManageInstallRuntime")}
-          </Button>
-        ) : null}
-      </div>
-    </div>
-  )
-}
-
-interface InstalledSkillsPaneProps {
-  cliUpdateError: string | null
-  cliVersionCheck: SkillVersionReport["cli"] | undefined
-  detailContentProps: SkillDetailContentProps
-  groups: ManagedSkillGroup[]
-  isExecutingCliUpdate: boolean
-  isDetailOpen: boolean
-  onCloseDetail: () => void
-  onSelectSkill: (skillId: string) => void
-  onUpdateCli: () => void
-  selectedSkill: ManagedSkillGroup | undefined
-  updateRegistrySkill: (skill: Pick<ManagedSkillGroup, "id" | "kind" | "packageName">) => void
-  updatingRegistrySkillId: string | null
-  versionCheckByKey: SkillVersionCheckByKey
-}
-
-function InstalledSkillsPane({
-  cliUpdateError,
-  cliVersionCheck,
-  detailContentProps,
-  groups,
-  isExecutingCliUpdate,
-  isDetailOpen,
-  onCloseDetail,
-  onSelectSkill,
-  onUpdateCli,
-  selectedSkill,
-  updateRegistrySkill,
-  updatingRegistrySkillId,
-  versionCheckByKey,
-}: InstalledSkillsPaneProps) {
-  const { t } = useAppI18n()
-
-  return (
-    <div className="min-h-0 overflow-auto px-3 py-3">
-      <div className="grid gap-3 pr-1">
-        <CliUpdateNotice
-          cli={cliVersionCheck}
-          error={cliUpdateError}
-          isUpdating={isExecutingCliUpdate}
-          onUpdate={onUpdateCli}
-        />
-        {groups.length === 0 ? (
-          <div className="oo-text-body oo-text-muted px-1 py-3">{t("skills.installedEmpty")}</div>
-        ) : (
-          <div className="grid grid-cols-[repeat(auto-fill,minmax(15.5rem,1fr))] gap-2.5">
-            {groups.map((group) => (
-              <InstalledSkillCard
-                key={group.id}
-                group={group}
-                selected={selectedSkill?.id === group.id}
-                updateRegistrySkill={updateRegistrySkill}
-                updatingRegistrySkillId={updatingRegistrySkillId}
-                versionCheck={getSkillVersionCheck(versionCheckByKey, group)}
-                onOpen={() => onSelectSkill(group.id)}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-
-      {isDetailOpen && selectedSkill ? (
-        <SkillManagementSheet title={selectedSkill.name} onClose={onCloseDetail}>
-          <SkillDetailContent {...detailContentProps} />
-        </SkillManagementSheet>
-      ) : null}
-    </div>
-  )
-}
-
-function CliUpdateNotice({
-  cli,
-  error,
-  isUpdating,
-  onUpdate,
-}: {
-  cli: SkillVersionReport["cli"] | undefined
-  error: string | null
-  isUpdating: boolean
-  onUpdate: () => void
-}) {
-  const { t } = useAppI18n()
-
-  if (cli?.status !== "update-available") {
-    return null
-  }
-
-  return (
-    <Card className="grid gap-2 rounded-md border-[var(--oo-warning-border)] bg-[var(--oo-warning-surface)] px-3 py-2 shadow-none">
-      <div className="flex min-w-0 items-center justify-between gap-3">
-        <div className="grid min-w-0 gap-1">
-          <div className="oo-text-label">{t("skills.cliUpdateAvailableTitle")}</div>
-          <CardDescription className="oo-text-caption-compact">
-            {t("skills.cliUpdateAvailableDescription", {
-              current: cli.currentVersion ?? t("skills.none"),
-              latest: cli.latestVersion ?? t("skills.none"),
-            })}
-          </CardDescription>
-        </div>
-        <Button type="button" variant="outline" size="sm" disabled={isUpdating} onClick={onUpdate}>
-          {isUpdating ? <AppIcons.status.loading className="animate-spin" /> : <AppIcons.action.download />}
-          {isUpdating ? t("skills.updatingCli") : t("skills.updateCli")}
-        </Button>
-      </div>
-      <SkillErrorNotice error={error} />
-    </Card>
-  )
-}
-
-interface InstalledSkillCardProps {
-  group: ManagedSkillGroup
-  onOpen: () => void
-  selected: boolean
-  updateRegistrySkill: (skill: Pick<ManagedSkillGroup, "id" | "kind" | "packageName">) => void
-  updatingRegistrySkillId: string | null
-  versionCheck: SkillVersionReport["skills"][number] | undefined
-}
-
-function InstalledSkillCard({
-  group,
-  onOpen,
-  selected,
-  updateRegistrySkill,
-  updatingRegistrySkillId,
-  versionCheck,
-}: InstalledSkillCardProps) {
-  const { t } = useAppI18n()
-  const status = getGroupStatus(group, t, getRuntimeHosts(group))
-  const hasUpdate = hasSkillUpdateAvailable(versionCheck)
-  const canUpdate = hasUpdate && shouldUpdatePublishedSkill(group)
-  const isPublishable = isPublishableLocalSkill(group)
-  const hasAttention = status.tone === "attention" || status.tone === "danger"
-  const statusLabel = hasUpdate
-    ? t("skills.updateAvailable")
-    : hasAttention
-      ? (status.label ?? t("skills.groupStatus.modified"))
-      : isPublishable
-        ? t("skills.publishable")
-        : t("skills.installed")
-  const badgeTone: ObjectStatusTone = hasUpdate ? "attention" : hasAttention ? status.tone : "ready"
-  const badgeClassName =
-    isPublishable && !hasUpdate && !hasAttention
-      ? publishableSkillBadgeClassName
-      : getSkillRowStatusBadgeClassName(badgeTone)
-  const packageLine = getGroupRowPackageLine(group) ?? getSkillKindLabel(group.kind, t)
-  const runtimeLabel =
-    hasUpdate && versionCheck
-      ? t("skills.versionUpdateAvailable", {
-          current: versionCheck.currentVersion ?? group.version ?? "",
-          latest: versionCheck.latestVersion ?? "",
-        })
-      : hasAttention
-        ? (status.description ?? t("skills.groupStatus.modifiedDescription", { count: 1 }))
-        : isPublishable
-          ? t("skills.publishableDescription")
-          : t("skills.installedDescription")
-  const isUpdating = updatingRegistrySkillId === group.id
-
-  return (
-    <div
-      className={cn(
-        "grid min-h-44 grid-rows-[minmax(0,1fr)_auto] overflow-hidden rounded-md border bg-card text-card-foreground transition-colors hover:bg-[var(--oo-row-hover)]",
-        selected && "border-[var(--accent-ring)] bg-[var(--oo-row-selected)] hover:bg-[var(--oo-row-selected)]",
-      )}
-    >
-      <button
-        type="button"
-        className="grid min-w-0 gap-2 p-3 text-left outline-none focus-visible:ring-[3px] focus-visible:ring-ring/40"
-        onClick={onOpen}
-      >
-        <div className="flex min-w-0 items-start gap-3">
-          <SkillIconFrame icon={group.icon} />
-          <div className="grid min-w-0 gap-1">
-            <div className="oo-text-label min-w-0 truncate">{group.name}</div>
-            <div className="oo-text-caption oo-text-muted min-w-0 truncate" title={packageLine}>
-              {packageLine}
-            </div>
-          </div>
-        </div>
-        {group.description ? (
-          <p className="oo-text-caption line-clamp-2 text-foreground/75">{group.description}</p>
-        ) : null}
-        <div className="oo-text-caption oo-text-muted min-w-0 truncate" title={runtimeLabel}>
-          {runtimeLabel}
-        </div>
-      </button>
-      <div className="oo-border-divider flex items-center justify-between gap-2 border-t px-3 py-2">
-        <Badge className={badgeClassName} variant={badgeTone === "danger" ? "destructive" : "outline"}>
-          {statusLabel}
-        </Badge>
-        {canUpdate ? (
-          <Button
-            type="button"
-            variant="outline"
-            size="sm"
-            disabled={isUpdating}
-            onClick={() => updateRegistrySkill(group)}
-          >
-            {isUpdating ? <AppIcons.status.loading className="animate-spin" /> : null}
-            {isUpdating ? t("skills.updatingRegistry") : t("skills.updateRegistry")}
-          </Button>
-        ) : (
-          <Button type="button" variant="ghost" size="sm" onClick={onOpen}>
-            {t("skills.installedManage")}
-          </Button>
-        )}
-      </div>
-    </div>
-  )
-}
-
-function SkillManagementSheet({
-  children,
-  onClose,
-  title,
-}: {
-  children: React.ReactNode
-  onClose: () => void
-  title: string
-}) {
-  const { t } = useAppI18n()
-  const sheetRef = React.useRef<HTMLElement | null>(null)
-
-  React.useEffect(() => {
-    const previousActiveElement = document.activeElement instanceof HTMLElement ? document.activeElement : null
-    const frame = window.requestAnimationFrame(() => {
-      sheetRef.current?.focus()
-    })
-
-    return () => {
-      window.cancelAnimationFrame(frame)
-      previousActiveElement?.focus()
-    }
-  }, [])
-
-  return (
-    <div
-      className="oo-modal-backdrop fixed inset-0 z-[120] [-webkit-app-region:no-drag]"
-      role="presentation"
-      onMouseDown={(event) => {
-        if (event.target === event.currentTarget) {
-          onClose()
-        }
-      }}
-    >
-      <aside
-        ref={sheetRef}
-        role="dialog"
-        aria-modal="true"
-        aria-label={title}
-        tabIndex={-1}
-        className="absolute top-0 right-0 grid h-full w-[min(30rem,calc(100vw-2rem))] grid-rows-[auto_minmax(0,1fr)] border-l bg-background shadow-xl [-webkit-app-region:no-drag]"
-        onKeyDown={(event) => {
-          if (event.key === "Escape") {
-            event.stopPropagation()
-            onClose()
-            return
-          }
-          if (event.key !== "Tab") {
-            return
-          }
-
-          const sheet = sheetRef.current
-          if (!sheet) {
-            return
-          }
-
-          const focusableElements = getFocusableElements(sheet)
-          if (focusableElements.length === 0) {
-            event.preventDefault()
-            sheet.focus()
-            return
-          }
-
-          const firstElement = focusableElements[0]
-          const lastElement = focusableElements[focusableElements.length - 1]
-          const activeElement = document.activeElement
-          if (event.shiftKey) {
-            if (activeElement === firstElement || activeElement === sheet || !sheet.contains(activeElement)) {
-              event.preventDefault()
-              lastElement.focus()
-            }
-            return
-          }
-
-          if (activeElement === lastElement || activeElement === sheet || !sheet.contains(activeElement)) {
-            event.preventDefault()
-            firstElement.focus()
-          }
-        }}
-        onMouseDown={(event) => event.stopPropagation()}
-      >
-        <div className="oo-border-divider flex min-w-0 items-center justify-between gap-3 border-b px-3 py-2 [-webkit-app-region:no-drag]">
-          <div className="oo-text-label min-w-0 truncate">{title}</div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon"
-            aria-label={t("skills.discoverCloseDetail")}
-            className="[-webkit-app-region:no-drag]"
-            onMouseDown={(event) => event.stopPropagation()}
-            onClick={onClose}
-          >
-            <AppIcons.action.cancel />
-          </Button>
-        </div>
-        <div className="min-h-0 overflow-auto p-3">{children}</div>
-      </aside>
-    </div>
-  )
-}
-
-function SkillIconFrame({
-  className,
-  icon,
-  iconClassName,
-}: {
-  className?: string
-  icon?: string
-  iconClassName?: string
-}) {
-  const normalizedIcon = normalizeSkillIconSource(icon)
-  const frameClassName = cn(
-    "flex size-10 shrink-0 items-center justify-center rounded-md border bg-background",
-    className,
-  )
-
-  if (isImageIcon(normalizedIcon)) {
-    return (
-      <span className={cn(frameClassName, "overflow-hidden")}>
-        <img alt="" src={normalizedIcon} className="size-full object-contain p-1.5" />
-      </span>
-    )
-  }
-
-  if (isEmojiIcon(normalizedIcon)) {
-    return <span className={cn(frameClassName, "text-xl")}>{normalizedIcon}</span>
-  }
-
-  return (
-    <span className={frameClassName}>
-      <SkillIcon icon={normalizedIcon} className={cn("size-5", iconClassName)} />
-    </span>
-  )
-}
-
-interface PublicSkillPackageSheetProps {
-  groupById: ManagedSkillGroupById
-  installingKey: string | null
-  locale: string
-  onClose: () => void
-  onInstall: (pkg: PublicSkillPackage, skillName?: string) => void
-  onOpenManagedSkill: (skillName: string) => void
-  pkg: PublicSkillPackage
-}
-
-function getFocusableElements(container: HTMLElement): HTMLElement[] {
-  return Array.from(
-    container.querySelectorAll<HTMLElement>(
-      [
-        "a[href]",
-        "button:not([disabled])",
-        "input:not([disabled])",
-        "select:not([disabled])",
-        "textarea:not([disabled])",
-        "[tabindex]:not([tabindex='-1'])",
-      ].join(","),
-    ),
-  ).filter((element) => !element.hasAttribute("disabled") && element.getAttribute("aria-hidden") !== "true")
-}
-
-function PublicSkillPackageSheet({
-  groupById,
-  installingKey,
-  locale,
-  onClose,
-  onInstall,
-  onOpenManagedSkill,
-  pkg,
-}: PublicSkillPackageSheetProps) {
-  const { t } = useAppI18n()
-  const sheetRef = React.useRef<HTMLElement | null>(null)
-
-  React.useEffect(() => {
-    const previousActiveElement = document.activeElement instanceof HTMLElement ? document.activeElement : null
-    const frame = window.requestAnimationFrame(() => {
-      sheetRef.current?.focus()
-    })
-
-    return () => {
-      window.cancelAnimationFrame(frame)
-      previousActiveElement?.focus()
-    }
-  }, [])
-
-  return (
-    <div
-      className="oo-modal-backdrop fixed inset-0 z-[120]"
-      role="presentation"
-      onMouseDown={(event) => {
-        if (event.target === event.currentTarget) {
-          onClose()
-        }
-      }}
-    >
-      <aside
-        ref={sheetRef}
-        role="dialog"
-        aria-modal="true"
-        aria-label={pkg.displayName}
-        tabIndex={-1}
-        className="absolute top-0 right-0 grid h-full w-[min(30rem,calc(100vw-2rem))] grid-rows-[auto_minmax(0,1fr)] border-l bg-background shadow-xl"
-        onKeyDown={(event) => {
-          if (event.key === "Escape") {
-            event.stopPropagation()
-            onClose()
-            return
-          }
-          if (event.key !== "Tab") {
-            return
-          }
-
-          const sheet = sheetRef.current
-          if (!sheet) {
-            return
-          }
-
-          const focusableElements = getFocusableElements(sheet)
-          if (focusableElements.length === 0) {
-            event.preventDefault()
-            sheet.focus()
-            return
-          }
-
-          const firstElement = focusableElements[0]
-          const lastElement = focusableElements[focusableElements.length - 1]
-          const activeElement = document.activeElement
-          if (event.shiftKey) {
-            if (activeElement === firstElement || activeElement === sheet || !sheet.contains(activeElement)) {
-              event.preventDefault()
-              lastElement.focus()
-            }
-            return
-          }
-
-          if (activeElement === lastElement || activeElement === sheet || !sheet.contains(activeElement)) {
-            event.preventDefault()
-            firstElement.focus()
-          }
-        }}
-        onMouseDown={(event) => event.stopPropagation()}
-      >
-        <div className="oo-border-divider flex min-w-0 items-center justify-between gap-3 border-b px-3 py-2">
-          <div className="oo-text-label min-w-0 truncate">{pkg.displayName}</div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon"
-            aria-label={t("skills.discoverCloseDetail")}
-            onClick={onClose}
-          >
-            <AppIcons.action.cancel />
-          </Button>
-        </div>
-        <div className="min-h-0 overflow-auto p-3">
-          <PublicSkillPackageDetail
-            groupById={groupById}
-            installingKey={installingKey}
-            locale={locale}
-            pkg={pkg}
-            onInstall={onInstall}
-            onOpenManagedSkill={onOpenManagedSkill}
-          />
-        </div>
-      </aside>
-    </div>
-  )
-}
-
-interface PublicSkillPackageDetailProps {
-  className?: string
-  groupById: ManagedSkillGroupById
-  installingKey: string | null
-  locale: string
-  onInstall: (pkg: PublicSkillPackage, skillName?: string) => void
-  onOpenManagedSkill: (skillName: string) => void
-  pkg: PublicSkillPackage
-}
-
-function PublicSkillPackageDetail({
-  className,
-  groupById,
-  installingKey,
-  locale,
-  onInstall,
-  onOpenManagedSkill,
-  pkg,
-}: PublicSkillPackageDetailProps) {
-  const { t } = useAppI18n()
-  const updateTime = formatPublicPackageUpdateTime(pkg.updateTime, locale)
-  const primarySkill = getPublicPackagePrimarySkill(pkg)
-  const primaryInstallSkill = getPublicPackagePrimaryInstallSkill(groupById, pkg)
-  const primaryState = getPublicPackageInstallState(groupById, pkg)
-  const isInstallingPrimary = installingKey === getPublicSkillInstallKey(pkg, primaryInstallSkill?.name)
-
-  return (
-    <aside className={cn("grid min-w-0 content-start gap-3", className)}>
-      <InspectorCard>
-        <CardHeader className="flex-row items-start gap-3 px-3 py-0">
-          <SkillIconFrame icon={pkg.icon} />
-          <div className="grid min-w-0 flex-1 gap-1">
-            <CardTitle className="oo-text-label min-w-0 truncate">{pkg.displayName}</CardTitle>
-            <CardDescription className="min-w-0 truncate">{pkg.name}</CardDescription>
-          </div>
-        </CardHeader>
-        <CardContent className="grid gap-2 px-3">
-          <div className="flex min-w-0 flex-wrap items-center gap-1">
-            <Badge variant="secondary">{pkg.version}</Badge>
-            {pkg.downloadCount === undefined ? null : (
-              <Badge variant="outline">{t("skills.discoverDownloads", { count: pkg.downloadCount })}</Badge>
-            )}
-            {updateTime ? <Badge variant="outline">{updateTime}</Badge> : null}
-          </div>
-          {pkg.description ? (
-            <CardDescription className="min-w-0 break-words text-foreground/80">{pkg.description}</CardDescription>
-          ) : null}
-          {primarySkill ? (
-            <div className="flex min-w-0 flex-wrap gap-1">
-              {primaryState === "installed" || primaryState === "name-conflict" ? (
-                <Button type="button" variant="outline" size="sm" onClick={() => onOpenManagedSkill(primarySkill.name)}>
-                  {t("skills.discoverOpenManage")}
-                </Button>
-              ) : (
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  disabled={isInstallingPrimary || !canInstallPublicSkill(primaryState)}
-                  onClick={() => onInstall(pkg, primaryInstallSkill?.name)}
-                >
-                  {isInstallingPrimary ? (
-                    <AppIcons.status.loading className="animate-spin" />
-                  ) : (
-                    <AppIcons.action.installPackage />
-                  )}
-                  {isInstallingPrimary
-                    ? t("skills.registryInstalling")
-                    : primaryState === "partially-installed"
-                      ? t("skills.discoverInstallMissing")
-                      : primaryState === "unavailable"
-                        ? t("skills.discoverUnavailable")
-                        : t("organizations.skillManageInstallRuntime")}
-                </Button>
-              )}
-            </div>
-          ) : null}
-        </CardContent>
-      </InspectorCard>
-
-      <InspectorInsetCard className="gap-2 px-3 py-2">
-        <div className="oo-text-caption-compact font-medium">{t("skills.discoverPackageInfo")}</div>
-        <div className="oo-text-caption-compact grid gap-1">
-          <div className="flex min-w-0 justify-between gap-3">
-            <span className="oo-text-muted">{t("skills.package")}</span>
-            <span className="min-w-0 truncate text-right">{pkg.name}</span>
-          </div>
-          <div className="flex min-w-0 justify-between gap-3">
-            <span className="oo-text-muted">{t("skills.discoverMaintainer")}</span>
-            <span className="min-w-0 truncate text-right">{getPublicPackageMaintainerLine(pkg, t)}</span>
-          </div>
-          {updateTime ? (
-            <div className="flex min-w-0 justify-between gap-3">
-              <span className="oo-text-muted">{t("skills.discoverUpdated")}</span>
-              <span className="min-w-0 truncate text-right">{updateTime}</span>
-            </div>
-          ) : null}
-        </div>
-      </InspectorInsetCard>
-    </aside>
-  )
 }
 
 interface SkillPeekProps {
